test cleanup: add deprecated marker, move benchmarks out
