+++ pytest.ini
@@ -11,3 +11,4 @@ log_file = logs/pytest.log
 markers =
     benchmark: mark test as benchmark
     slow: mark test as slow
+    deprecated: mark test as covering a deprecated feature
